+++ NINA.Plugin.TargetScheduler/Dockables/ImagingDockableTemplate.xaml.cs
@@ -1,0 +1,13 @@
+using System.ComponentModel.Composition;
+using System.Windows;
+
+namespace NINA.Plugin.TargetScheduler.Dockables {
+
+    [Export(typeof(ResourceDictionary))]
+    public partial class ImagingDockableTemplate {
+
+        public ImagingDockableTemplate() {
+            InitializeComponent();
+        }
+    }
+}