+++ NINA.Plugin.TargetScheduler/PubSub/NewTargetStartPublisher.cs
@@ -1,15 +1,17 @@
 ﻿using NINA.Plugin.Interfaces;
 using NINA.Plugin.TargetScheduler.Planning;
+using NINA.Plugin.TargetScheduler.Planning.Interfaces;
 
 namespace NINA.Plugin.TargetScheduler.PubSub {
 
     public class NewTargetStartPublisher : TSPublisher {
+        public const string TOPIC = "TargetScheduler-NewTargetStart";
 
         public NewTargetStartPublisher(IMessageBroker messageBroker) : base(messageBroker) {
         }
 
-        public override string Topic => "TargetScheduler-NewTargetStart";
-        public override int Version => 1;
+        public override string Topic => TOPIC;
+        public override int Version => 2;
 
         public void Publish(SchedulerPlan plan) {
             Publish(GetMessage(plan));
@@ -19,8 +21,17 @@ namespace NINA.Plugin.TargetScheduler.PubSub {
             TSMessage message = new TSMessage(Topic, plan.PlanTarget.Name, MessageSender, MessageSenderId, Version);
             message.Expiration = plan.PlanTarget.EndTime;
             message.CustomHeaders.Add("ProjectName", plan.PlanTarget.Project.Name);
+            message.CustomHeaders.Add("TargetName", plan.PlanTarget.Name);
             message.CustomHeaders.Add("Coordinates", plan.PlanTarget.Coordinates);
             message.CustomHeaders.Add("Rotation", plan.PlanTarget.Rotation);
+
+            IExposure exp = plan.PlanTarget.SelectedExposure;
+            message.CustomHeaders.Add("ExposureFilterName", exp.FilterName);
+            message.CustomHeaders.Add("ExposureLength", exp.ExposureLength);
+            message.CustomHeaders.Add("ExposureGain", exp.Gain.HasValue ? exp.Gain.ToString() : "(camera)");
+            message.CustomHeaders.Add("ExposureOffset", exp.Offset.HasValue ? exp.Offset.ToString() : "(camera)");
+            message.CustomHeaders.Add("ExposureBinning", exp.BinningMode.ToString());
+
             return message;
         }
     }